pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opensearch-fscrawler"
version = "2.10.0"
description = "File system crawler for OpenSearch / Elasticsearch — Python rewrite of FSCrawler"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
authors = [
    { name = "Toomas Ormisson" },
]
maintainers = [
    { name = "Toomas Ormisson" },
]
keywords = ["elasticsearch", "opensearch", "crawler", "indexer", "tika", "fscrawler"]
dependencies = [
    "opensearch-py>=2.4",
    "pyyaml>=6.0",
    "click>=8.1",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "httpx>=0.27",
    "python-dateutil>=2.9",
    "watchdog>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/P6rguVyrst/opensearch-fscrawler"
"Upstream (Java)" = "https://github.com/dadoonet/fscrawler"
"Bug Tracker" = "https://github.com/P6rguVyrst/opensearch-fscrawler/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
    "mypy>=1.9",
    "types-PyYAML>=6.0",
    "types-python-dateutil>=2.9",
    "responses>=0.25",
]

[project.scripts]
fscrawler = "fscrawler.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
addopts = "-sv --cov=fscrawler --cov-report=term-missing --cov-report=html:htmlcov"

[tool.coverage.run]
source = ["src/fscrawler"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
